% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_interval.R
\name{ctrl_fit_recipe}
\alias{ctrl_fit_recipe}
\title{Control function used as part of \code{prep_interval()}
could likely be improved with newer extracting funs in tidymodels}
\usage{
ctrl_fit_recipe(x)
}
\description{
Control function used as part of \code{prep_interval()}
could likely be improved with newer extracting funs in tidymodels
}
