% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distGeoAni.R
\name{distGeoAni}
\alias{distGeoAni}
\title{Compute anistropy corrected distance between two sets of data}
\usage{
distGeoAni(xrow, yrow, xcol, ycol, rotate = 0, range = 1, minorp = 1)
}
\arguments{
\item{xrow}{vector with x-coordinates that will form rows of distance matrix}

\item{yrow}{vector with y-coordinates that will form rows of distance matrix, must be of same length as xrow}

\item{xcol}{vector with x-coordinates that will form columns of distance matrix}

\item{ycol}{vector with y-coordinates that will form columns of distance matrix, must be of same length as xcol}

\item{rotate}{rotation of anisotropic axes, default = 0}

\item{range}{range of autocorrelation model, default = 1}

\item{minorp}{proportion of range in x direction to that of y direction for unrotated anisotropic model, default = 1}
}
\value{
matrix of distances
}
\description{
computes anistropy corrected distance between two sets of data
}
\author{
Jay Ver Hoef
}
