% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fefe.R
\name{fefe}
\alias{fefe}
\title{Fixed effects estimation for geostatistical linear model}
\usage{
fefe(cope_out = NULL, data, spatial_model = corModelExponential,
  theta = NULL, z = NULL, X = NULL, xcoords = NULL,
  ycoords = NULL, extrap = NULL, par = FALSE, subsampindx = NULL,
  compute_covbalt1 = FALSE, compute_covbalt2 = FALSE)
}
\arguments{
\item{cope_out}{Object of class 'cope.'  Output from cope() function.}

\item{data}{Data set used with cope() function}

\item{spatial_model}{Spatial autocorrelation models for errors.  The list of spatial autocorrelation 
models is "exponential","expRadon2","expRadon4","gaussian","stable",
"rationalQuad","cauchyGrav","cauchyMag","cauchy","circular","spherical",
"cubic","penta","cardinalSine","besselK","besselJ"  Default is "exponential".
Any names in the list not given above will be searched among the columns in the data set and
used as a factor variable for levels of a traditional random effect.}

\item{theta}{input the covariance parameters directly}

\item{z}{a vector of observed data}

\item{X}{design matrix for observed data}

\item{xcoords}{x-coordinates for observed data.  Same dimension as z.}

\item{ycoords}{y-coordinates for observed data.  Same dimension as z.}

\item{extrap}{extra parameter for those autocorrelation models that use it.
Currently, these are models "stable", "besselK", "besselJ".}

\item{subsampindx}{Column containing indexes for data partitioning.}

\item{use}{parallel processing?  TRUE or FALSE.}
}
\value{
a list of class "splmm".  The functions "summary" and "print" are used to obtain and print a summary. "anova" returns just the analysis of variance table...
}
\description{
Fixed effects estimation for geostatistical linear model
}
\author{
Jay Ver Hoef
}
