% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2LLgPAR.R
\name{m2LLgPAR}
\alias{m2LLgPAR}
\title{minus 2 times loglikelihood for case 12}
\usage{
m2LLgPAR(theta, z, X, xcoords, ycoords, estMeth, grpindx)
}
\arguments{
\item{theta}{vector of estimated covariance parameters}

\item{z}{vector of data}

\item{X}{design matrix for fixed effects}

\item{xcoords}{vector with the x-coordinates}

\item{ycoords}{vector with the y-coordinates}

\item{estMeth}{estimation method.  Default is "REML" for restricted maximum likelihood.  Other options are "ML" for maximum likelihood}

\item{Z}{list of design matrices for each random effect}
}
\value{
minus 2 times the loglikelihood
}
\description{
fits a geostatistical linear model
}
\author{
Jay Ver Hoef
}
