% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass.R
\name{mass}
\alias{mass}
\title{Makes subsamples for faster model fitting}
\usage{
mass(x, y, ngroups, gsize = NULL, ssmeth)
}
\arguments{
\item{x}{vector of x-coordinates}

\item{y}{vector of y-coordinates}

\item{ngroups}{number of groups to make}

\item{gsize}{group size, only used for 'subsemble' method}

\item{ssmeth}{subsampling method.  Valid methods are 'random' for 
randomly assigning locations to groups, 'compKmean' for using K-means
on coordinates to create compact spatial groups, 'zimmer' for 
using K-means to create compact spatial groups, but then randomly
re-assigning 10 percent in each group to get more distant distances,
'subsemble' for ngroups of size gsize as in Barbian and Assuncao
(2017).}
}
\value{
a matrix where the first column is the row index of the x
(and y) vector, and the second column is group membership.
}
\description{
Makes subsamples for faster model fitting
}
\author{
Jay Ver Hoef
}
