% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pints-package.R
\docType{package}
\name{pints-package}
\alias{pints}
\alias{pints-package}
\title{pints: Prediction Intervals Simulated}
\description{
To learn more about pints see https://github.com/brshallo/pints
}
\author{
\strong{Maintainer}: Bryan Shalloway \email{brshallodev@gmail.com}

}
\keyword{internal}
