% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poba.R
\name{poba}
\alias{poba}
\title{Prediction over block aggregrations for geostatistical models}
\usage{
poba(theta, z, X, Xp, xcoords, ycoords, xcoordsp, ycoordsp,
  predmeth = "alldata", nNN = 50, par = FALSE)
}
\arguments{
\item{theta}{covariance parameters to be used in estimating fixed effects}

\item{z}{a vector of observed data}

\item{X}{design matrix for observed data}

\item{Xp}{design matrix for prediction data}

\item{xcoords}{x-coordinates for observed data.  Same dimension as z.}

\item{ycoords}{y-coordinates for observed data.  Same dimension as z.}

\item{xcoordsp}{x-coordinates for prediction data.  Same number of rows as Xp.}

\item{ycoordsp}{y-coordinates for prediction data.  Same number of rows as Xp.}

\item{predmeth}{prediction method.  'alldata' uses all data. 'NNdata' uses nearest neighbor method.}

\item{nNN}{number of nearest neighbors to use when predmeth='NNdata'.}

\item{use}{parallel processing?  TRUE or FALSE.}
}
\value{
a list of two scalars; one is the block prediction, the other the standard error for that prediction.
}
\description{
Prediction over block aggregrations for geostatistical models
}
\author{
Jay Ver Hoef
}
