% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_interval.R
\name{predict_interval}
\alias{predict_interval}
\title{Predict Interval}
\usage{
predict_interval(
  prepped_interval,
  new_data,
  probs = c(0.025, 0.5, 0.975),
  cross = TRUE,
  n_sims = 10000
)
}
\arguments{
\item{prepped_interval}{Object outputted by \code{prep_interval()}.}

\item{new_data}{Data to generate predictions on.}

\item{probs}{Quantiles of predictions to output.}

\item{cross}{If TRUE makes distribution for selecting quantiles from made up
of all possible combinations of samples of {model fitting uncertainty} and
{sample uncertainty}. If FALSE, create \code{n_sims} number of simulations for
each sample.}

\item{n_sims}{An integer of length 1, specifying the number of simulations for each observation. Default of \code{cross == TRUE},
means \code{n_sims} is ignored. If \code{cross == FALSE}, default of \code{n_sims} is 10000.}
}
\value{
A tibble containing columns \verb{probs_*} at each quantile specified by \code{probs}.
}
\description{
This function takes in the output from \code{prep_interval()} along with an
unprepped hold-out dataset and outputs a prediction interval.
}
\examples{
# See README
}
