% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_interval.R
\name{prep_interval}
\alias{prep_interval}
\title{Prep for Prediction Interval}
\usage{
prep_interval(wf, fit_data, n_boot = sqrt(nrow(fit_data)))
}
\arguments{
\item{wf}{Tidymodels workflow containing a recipe and a model.}

\item{fit_data}{Dataframe used to estimate the prediction intervals.}

\item{n_boot}{Number of bootstrapped samples used to simulate building the
model (used to determine uncertainty due to model).}
}
\value{
A named list of two tibbles named \code{model_uncertainty} and
\code{sample_uncertainty}.
}
\description{
This function takes in a workflow and outputs a named list meant to be passed
into \code{predict_interval()}.
}
\examples{
# See README
}
