% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulo.R
\name{pulo}
\alias{pulo}
\title{Predict Method for Spatial Linear Mixed Model Fits}
\usage{
pulo(theta, z, X, Xp, xcoords, ycoords, xcoordsp, ycoordsp,
  predmeth = "nearnei", nNN = 50, par = FALSE)
}
\arguments{
\item{theta}{covariance parameters to be used in estimating fixed effects}

\item{z}{a vector of observed data}

\item{X}{design matrix for observed data}

\item{Xp}{design matrix for prediction data}

\item{xcoords}{x-coordinates for observed data.  Same dimension as z.}

\item{ycoords}{y-coordinates for observed data.  Same dimension as z.}

\item{xcoordsp}{x-coordinates for prediction data.  Same number of rows as Xp.}

\item{ycoordsp}{y-coordinates for prediction data.  Same number of rows as Xp.}

\item{predmeth}{prediction method.  'claskrig' uses all data. 'nearnei' uses nearest neighbor method.}

\item{nNN}{number of nearest neighbors to use when predmeth='nearnei'.}

\item{use}{parallel processing?  TRUE or FALSE.}
}
\value{
\code{predict.splm} produces a data.frame with 2 columns for predictions, and prediction standard errors for missing response variable, one with the response variable name appended by \code{.pred} for predictions and by \code{.predSE} for the prediction standard errors.
}
\description{
Makes predictions for missing data or using external prediction data set.
}
\details{
These are the universal kriging equations found in Cressie (1993, pg. 154-155).
}
\references{
\cite{Cressie, N.A.C. (1993) Statistics for Spatial Data. Wiley.}
}
\author{
Jay Ver Hoef
}
