% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim2DSinSurf.R
\name{sim2DSinSurf}
\alias{sim2DSinSurf}
\title{simulated large spatial data sets from combination of sine wave surfaces with random amplitudes, freqencies, and spatial shifts.}
\usage{
sim2DSinSurf(x, y)
}
\arguments{
\item{x}{a vector of x-coordinates}

\item{y}{a vector of y-coordinates}
}
\value{
data.frame of simulated surface (z) column.
}
\description{
simulated large spatial data sets from combination of sine wave surfaces with random amplitudes, freqencies, and spatial shifts.
}
\author{
Jay Ver Hoef
}
