% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spin-package.R
\docType{package}
\name{spin-package}
\alias{spin}
\alias{spin-package}
\title{spin: Simulating Prediction Intervals}
\description{
To learn more about spin see https://github.com/brshallo/spin
}
\author{
\strong{Maintainer}: Bryan Shalloway \email{brshallodev@gmail.com}

}
\keyword{internal}
