\name{SPIn-package}
\alias{SPIn-package}
\docType{package}
\title{
Optimal Shortest Probability Intervals
}
\description{
Adopt an optimal weighting strategy to compute better shortest probability intervals (SPIn).
}
\details{
\tabular{ll}{
Package: \tab SPIn\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-12-30\cr
License: \tab GPL (>= 2)\cr
}
This package contains functions for constructing and plotting optimal shortest probability intervals.
}
\author{
Ying Liu

Maintainer: Ying Liu <yliu@stat.columbia.edu>
}
\references{
An Optimal Weighting Strategy for Better Shortest Probability Intervals. (In preparation)
}
