\name{SPIn-package}
\alias{SPIn-package}
\docType{package}
\title{
Simulation Efficient Shortest Probability Intervals
}
\description{
Implement an optimal weighting strategy to compute simulation efficient shortest probability intervals (spin's).
}
\details{
\tabular{ll}{
Package: \tab SPIn\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2013-4-2\cr
License: \tab GPL (>= 2)\cr
}
This package contains functions for constructing and plotting simulation efficient shortest probability intervals.
}
\author{
Ying Liu

Maintainer: Ying Liu <yliu@stat.columbia.edu>
}
\references{
Simulation efficient shortest probability intervals. (arXiv:1302.2142)
}
